Use %lld to fomat a2_gid_t

+++ src/RequestGroup.h
@@ -73,7 +73,8 @@ class BtRuntime;
 class PeerStorage;
 #endif // ENABLE_BITTORRENT
 
-typedef int64_t a2_gid_t;
+// To make %lld happy, we use long long int instead of int64_t.
+typedef long long int a2_gid_t;
 
 class RequestGroup {
 public:
